test parsers

+++ mlflow_fastapi/prepare_model.py
@@ -52,9 +52,15 @@ PREDICTIONS_WRAPPER_ATTR_NAME_ENV_KEY = "PREDICTIONS_WRAPPER_ATTR_NAME"
 _logger = logging.getLogger(__name__)
 
 # model_uri = "models:/ElasticnetWineModel/Staging"
-def do_prepare_model(model_uri):
-  model_path = os.getcwd() + "/model"
-  shutil.rmtree(model_path)
 
-  local_path = _download_artifact_from_uri(model_uri)
-  shutil.copytree(local_path, model_path)
+
+def do_prepare_model(model_uri):
+    model_path = os.getcwd() + "/model"
+    # home_path = os.environ['MLFLOW_FASTAPI_HOME']
+    # model_path = home_path + "/model"
+
+    if os.path.exists(model_path):
+        shutil.rmtree(model_path)
+
+    local_path = _download_artifact_from_uri(model_uri)
+    shutil.copytree(local_path, model_path)
